Split the second half of camera.c to controller.c. This is a fully matched and documented file!
