+++ src/game_text.c
@@ -6,6 +6,7 @@
 #include "objects.h"
 #include "racer.h"
 #include "game.h"
+#include "controller.h"
 
 /************ .data ************/
