+++ src/menu.c
@@ -30,6 +30,7 @@
 #include "unknown_078050.h"
 #include "unknown_0255E0.h"
 #include "lib/src/mips1/al/alSynStartVoiceParams.h"
+#include "controller.h"
 
 /**
  * @file Contains all the code used for every menu in the game.
