+++ src/racer.c
@@ -26,6 +26,7 @@
 #include "object_models.h"
 #include "audiosfx.h"
 #include "collision.h"
+#include "controller.h"
 
 
 #define MAX_NUMBER_OF_GHOST_NODES 360
